Increase the width of the grounded pins bridges and update DRC
The connections between the GND pins and the GND plane were too thin, so on the milling process they ended up not connected.

To solve this problem the width of those bridges were increased from 0.1016mm to 0.762mm changinf the GND polygon to 1.016mm width.

A new DRC was created as is used by both boards.